refactor: Organize downloads screen components
- Group `ArchivedScreen` with the downloads screen feature
- Colocate `DownloadPrimaryActionButton` with download item UI
- Extract list content and swipe interactions into focused composables
